requirements.txt
accelerate==1.13.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair @ file:///home/conda/feedstock_root/build_artifacts/altair-split_1763475854573/work
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
backports.zstd @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_backports.zstd_1767045020/work
beautifulsoup4==4.14.3
blinker @ file:///home/conda/feedstock_root/build_artifacts/blinker_1731096409132/work
blis==1.3.3
Brotli @ file:///Users/runner/miniforge3/conda-bld/brotli-split_1764017116075/work
cachetools @ file:///home/conda/feedstock_root/build_artifacts/cachetools_1773120366885/work
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_click_1764518652/work
cloudpathlib==0.23.0
confection==0.1.5
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
cymem==2.0.13
dataclasses-json==0.6.7
Deprecated==1.3.1
distro==1.9.0
emoji==2.15.0
en_core_web_sm==3.8.0
faiss==1.9.0
faiss-cpu==1.13.2
filelock==3.25.2
filetype==1.2.0
flatbuffers==25.12.19
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
gitdb @ file:///home/conda/feedstock_root/build_artifacts/gitdb_1735887193964/work
GitPython @ file:///home/conda/feedstock_root/build_artifacts/gitpython_1767358723114/work
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.192.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-cloud-vision==3.12.1
google-genai==1.67.0
google-generativeai==0.8.6
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.16.0
h2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_h2_1756364871/work
hf-xet==1.4.2
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.7.1
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna==3.11
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
installer==0.7.0
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_jinja2_1764517220/work
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_jsonschema_1767839954/work
jsonschema-specifications @ file:///tmp/tmptzik3dqa/src
kiwisolver==1.5.0
langchain==1.2.12
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.19
langchain-google-genai==4.2.1
langchain-ollama==1.0.1
langchain-openai==1.1.11
langchain-text-splitters==1.1.1
langdetect==1.0.9
langgraph==1.1.2
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.11
langsmith==0.7.17
llvmlite==0.46.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe @ file:///Users/runner/miniforge3/conda-bld/markupsafe_1772445084013/work
marshmallow==3.26.2
matplotlib==3.10.8
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
multidict==6.7.1
murmurhash==1.0.15
mypy_extensions==1.1.0
narwhals @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_narwhals_1773164865/work
networkx==3.6.1
numba==0.64.0
numpy==2.4.3
olefile==0.47
ollama==0.6.1
onnx==1.20.1
onnxruntime==1.24.3
openai==2.28.0
opencv-python==4.13.0.92
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas @ file:///Users/runner/miniforge3/conda-bld/pandas_1764614611550/work
pdf2image @ file:///home/conda/feedstock_root/build_artifacts/pdf2image_1733174111200/work
pdfminer.six==20251230
pdfplumber==0.11.9
pi_heif==1.3.0
pikepdf==10.5.0
pillow @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_pillow_1770794152/work
preshed==3.0.12
propcache==0.4.1
proto-plus==1.27.1
protobuf==5.29.6
psutil==7.2.2
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck @ file:///home/conda/feedstock_root/build_artifacts/pydeck_1738346561094/work
Pygments==2.19.2
pyparsing==3.3.2
pypdf==6.8.0
pypdfium2==5.6.0
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
pytesseract @ file:///home/conda/feedstock_root/build_artifacts/pytesseract_1736249287101/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
python-dotenv==1.2.2
python-iso639==2026.1.31
python-magic==0.4.27
python-multipart==0.0.22
python-oxmsg==0.0.2
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1742920838005/work
PyYAML==6.0.3
RapidFuzz==3.14.3
referencing @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_referencing_1760379115/work
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_rpds-py_1764543142/work
safetensors==0.7.0
scipy==1.17.1
shellingham==1.5.4
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
smart_open==7.5.1
smmap @ file:///home/conda/feedstock_root/build_artifacts/smmap_1739781697784/work
sniffio==1.3.1
soupsieve==2.8.3
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.48
srsly==2.5.2
streamlit @ file:///home/conda/feedstock_root/build_artifacts/streamlit_1772724930433/work
sympy==1.14.0
tenacity==9.1.4
thinc==8.3.10
tiktoken==0.12.0
timm==1.0.25
tokenizers==0.22.2
toml @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_toml_1764486833/work
torch==2.10.0
torchvision==0.25.0
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1765836484236/work
tqdm==4.67.3
transformers==5.3.0
typer==0.24.1
typer-slim==0.24.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1765719872007/work
unstructured==0.21.5
unstructured-client==0.42.10
unstructured.pytesseract==0.3.15
unstructured_inference==1.2.0
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.1
wasabi==1.1.3
watchdog @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_watchdog_1772608105/work
weasel==0.4.3
webencodings==0.5.1
websockets==16.0
wrapt==1.17.3
xxhash==3.6.0
yarl==1.23.0
zipp @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_zipp_1764460141/work
zstandard==0.25.0
